backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
scikit-learn
numpy
scipy
rapidfuzz